requirements.txt
asgiref==3.5.1
Django==4.0.4
sqlparse==0.4.2
tzdata==2022.1
django-allauth==0.49.0
virtualenv==20.13.1
djangorestframework==3.13.1